+++ CHANGELOG.md
@@ -3,7 +3,9 @@ Changelog
 
 ## v0.2.0 (2023-10-26)
 
- * Add `RacistJester` and [`CODE_OF_CONDUCT`](./CODE_OF_CONDUCT.md)
+  * Added [`CODE_OF_CONDUCT.md`](./CODE_OF_CONDUCT)
+  * Added `RacistJester` (`--racist`)
+  * Added `OfficeJester` (`--office`, `--twss`)
 
 ## v0.1.0 (2023-10-25)